Addressing issue #8 by adding package level comments to the search.mb package and cleaning up the code a little. Moved a couple of unit tests to the test package.
