+++ src/main/java/edu/cmu/tetrad/search/mb/Mmhc.java
@@ -36,9 +36,7 @@ import java.util.List;
 import java.util.Map;
 
 /**
- * Implements the PC ("Peter/Clark") algorithm, as specified in Chapter 6 of Spirtes, Glymour, and Scheines, "Causation,
- * Prediction, and Search," 2nd edition, with a modified rule set in step D due to Chris Meek. For the modified rule
- * set, see Chris Meek (1995), "Causal inference and causal explanation with background knowledge."
+ * Implements the MMHC algorithm.
  *
  * @author Joseph Ramsey (this version).
  */
